+++ exportplan/serializers.py
@@ -1,3 +1,4 @@
+import datetime
 import decimal
 import json
 
@@ -81,9 +82,25 @@ class UiOptions(serializers.Serializer):
         return value[0].split(',')
 
 
+class UiProgress(serializers.Serializer):
+    is_complete = serializers.BooleanField(required=False)
+    date_last_visited = serializers.CharField(required=False)
+    modified = serializers.DateTimeField(required=False)
+
+    def to_representation(self, instance):
+        modified = instance.get('modified')
+        if isinstance(modified, (datetime.date, datetime.datetime)):
+            instance['date_last_visited'] = modified.isoformat()
+            instance.pop('modified')
+        return instance
+
+    class Meta:
+        exclude = ('ts',)
+
+
 class FundingAndCreditSerializer(serializers.Serializer):
-    override_estimated_total_cost = serializers.FloatField(required=True)
-    funding_amount_required = serializers.FloatField(required=True)
+    override_estimated_total_cost = serializers.FloatField(required=False)
+    funding_amount_required = serializers.FloatField(required=False)
 
 
 class DirectCostsSerializer(serializers.Serializer):
@@ -174,6 +191,16 @@ class TotalCostAndPriceSerializer(serializers.Serializer):
         return potential_total_profit
 
 
+class GettingPaidSerializer(serializers.Serializer):
+    class PaymentMethodSerializer(serializers.Serializer):
+        method = serializers.ListField(child=serializers.CharField(), required=False)
+        notes = serializers.CharField(required=False, allow_blank=True, validators=[no_html])
+
+    payment_method = PaymentMethodSerializer(required=False)
+    payment_terms = PaymentMethodSerializer(required=False)
+    incoterms = PaymentMethodSerializer(required=False)
+
+
 class ExportPlanSerializer(serializers.Serializer):
     export_commodity_codes = ExportPlanCommodityCodeSerializer(many=True, required=False)
     export_countries = ExportPlanCountrySerializer(many=True, required=False)
@@ -184,10 +211,18 @@ class ExportPlanSerializer(serializers.Serializer):
     marketing_approach = MarketingApproachSerializer(required=False)
     adaptation_target_market = AdaptationTargetMarketSerializer(required=False)
     ui_options = UiOptions(required=False)
+    ui_progress = serializers.DictField(child=UiProgress(), required=False)
     direct_costs = DirectCostsSerializer(required=False)
     overhead_costs = OverheadCostsSerializer(required=False)
     total_cost_and_price = TotalCostAndPriceSerializer(required=False)
     funding_and_credit = FundingAndCreditSerializer(required=False)
+    getting_paid = GettingPaidSerializer(required=False)
+
+    def to_internal_value(self, data):
+        internal_val = super().to_internal_value(data)
+        if internal_val.get('ui_progress') == {}:
+            internal_val.pop('ui_progress')
+        return internal_val
 
     @property
     def total_direct_costs(self):