+++ react-components/src/components/FundingCredit/FundingCreditOptions/Total.jsx
@@ -1,0 +1,32 @@
+import React, { memo } from 'react'
+import PropTypes from 'prop-types'
+
+export const Total = memo(({ total, currency, label }) => {
+  return (
+    <div className="costs costs--total bg-blue-deep-80 text-white">
+      <table className="m-t-0">
+        <tbody>
+          <tr>
+            <td>{label}</td>
+            <td className="total">
+              <span className="text-white m-r-s">{currency}</span>
+              <span className="body-l-b text-white">{total}</span>
+            </td>
+          </tr>
+        </tbody>
+      </table>
+    </div>
+  )
+})
+
+Total.propTypes = {
+  total: PropTypes.number,
+  currency: PropTypes.string,
+  label: PropTypes.string,
+}
+
+Total.defaultProps = {
+  total: 0,
+  currency: 'GBP',
+  label: 'Total funding',
+}